# Optimized `makeLookupFunction`.