I've noticed that the `omitUnnecessaryFields` function pops up more and more often in our profiling traces, sometimes taking even 26% of the CPU time. It's clear that it does the complete opposite of what all JIT languages like - it alters the created structs. I replaced it with a `buildResult` function that creates an object dynamically, only with the properties that should be there.

This improvement made up roughly 70% of the results. But then I made the `arrayIndices` calculation lazy, and while it didn't change anything for Safari, it got me another significant improvement on V8.

<details>
<summary>Benchmarks</summary>

```ts
{
  const matcher = MinimongoTest.makeLookupFunction('a');
  const doc = { a: 1 };
  for (let round = 1; round <= 50; ++round) {
    console.time('simple');
    for (let _ = 0; _ < 10000; ++_) matcher(doc);
    console.timeEnd('simple');
  }
}

{
  const matcher = MinimongoTest.makeLookupFunction('a');
  const doc = { a: { b: { c: { d: { e: 1 } } } } };
  for (let round = 1; round <= 50; ++round) {
    console.time('complex');
    for (let _ = 0; _ < 10000; ++_) matcher(doc);
    console.timeEnd('complex');
  }
}

{
  const matcher = MinimongoTest.makeLookupFunction('a.b.c');
  const doc = { a: [{ b: [{ c: 1 }] }] };
  for (let round = 1; round <= 50; ++round) {
    console.time('arrays');
    for (let _ = 0; _ < 10000; ++_) matcher(doc);
    console.timeEnd('arrays');
  }
}
```

Aggregated using https://vazco.github.io/console-time-analyzer/.

Node 14.21.2, before:
|  label  |    avg    |    max    |    min    |    std    |    sum    |  n  |
|:-------:|:---------:|:---------:|:---------:|:---------:|:---------:|:---:|
| simple  | 1.50276ms | 5.46500ms | 1.26900ms | 0.64776ms | 150.276ms | 100 |
| complex | 1.39677ms | 3.63200ms | 1.26100ms | 0.28840ms | 139.677ms | 100 |
| arrays  | 6.43423ms | 13.0920ms | 5.98600ms | 0.93085ms | 643.423ms | 100 |

Node 14.21.2, after:
|  label  |    avg    |    max    |    min    |    std    |    sum    |  n  |
|:-------:|:---------:|:---------:|:---------:|:---------:|:---------:|:---:|
| simple  | 0.08248ms | 3.12100ms | 0.02200ms | 0.34443ms | 8.24800ms | 100 |
| complex | 0.08322ms | 2.81100ms | 0.02200ms | 0.31567ms | 8.32200ms | 100 |
| arrays  | 3.27812ms | 9.01800ms | 2.95700ms | 0.83409ms | 327.812ms | 100 |

Chrome 109, before:
|  label  |    avg    |    max    |    min    |    std    |    sum    |  n  |
|:-------:|:---------:|:---------:|:---------:|:---------:|:---------:|:---:|
| simple  | 1.20268ms | 2.90600ms | 1.08886ms | 0.20236ms | 120.268ms | 100 |
| complex | 1.17900ms | 1.44677ms | 1.09887ms | 0.06908ms | 117.900ms | 100 |
| arrays  | 5.46177ms | 7.32226ms | 5.26611ms | 0.23321ms | 546.177ms | 100 |

Chrome 109, after:
|  label  |    avg    |    max    |    min    |    std    |    sum    |  n  |
|:-------:|:---------:|:---------:|:---------:|:---------:|:---------:|:---:|
| simple  | 0.14662ms | 1.12402ms | 0.07592ms | 0.13918ms | 14.2221ms | 100 |
| complex | 0.12575ms | 0.27392ms | 0.09301ms | 0.04746ms | 12.4499ms | 100 |
| arrays  | 3.18474ms | 6.59887ms | 2.88696ms | 0.49701ms | 315.289ms | 100 |

Safari 16.2, before:
|  label  |    avg    |    max    |    min    |    std    |    sum    |  n  |
|:-------:|:---------:|:---------:|:---------:|:---------:|:---------:|:---:|
| simple  | 0.74933ms | 3.60500ms | 0.58200ms | 0.37448ms | 74.1840ms | 100 |
| complex | 0.71135ms | 2.17000ms | 0.58600ms | 0.27525ms | 71.1350ms | 100 |
| arrays  | 7.84627ms | 9.23800ms | 6.97600ms | 0.62878ms | 784.627ms | 100 |

Safari 16.2, after:
|  label  |    avg    |    max    |    min    |    std    |    sum    |  n  |
|:-------:|:---------:|:---------:|:---------:|:---------:|:---------:|:---:|
| simple  | 0.85201ms | 9.60500ms | 0.65200ms | 0.91024ms | 83.4970ms | 100 |
| complex | 0.76904ms | 1.76400ms | 0.64800ms | 0.20524ms | 76.1350ms | 100 |
| arrays  | 6.92365ms | 9.40700ms | 6.00600ms | 0.53365ms | 692.365ms | 100 |

</details>

As you can see, the gains are _massive_ (50%-90%) for V8 (Node and Chrome) while Safari is slightly slower (10%) in simpler cases and slightly faster (10%) for arrays.

I started of the latest 2.10 release, but it could be backported a long way back.